requirements.txt
requests
bs4
lxml
python-dotenv
schedule
apscheduler
